refactor: moved NamedList component from utils/ to list/ and change utils directory in components.json from utils/ to /utils/shadcn-helper
