+++ package.json
@@ -34,7 +34,7 @@
         "@vitest/coverage-v8": "^2.1.8",
         "@vitest/ui": "^2.1.5",
         "autoprefixer": "^10.4.20",
-        "bits-ui": "^1.0.0-next.71",
+        "bits-ui": "^1.0.0-next.74",
         "clsx": "^2.1.1",
         "eslint": "^9.17.0",
         "eslint-config-prettier": "^9.1.0",
