+++ src/lib/components/primitives/avatar/avatar-image.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { Avatar as AvatarPrimitive } from "bits-ui";
-    import { cn } from "$lib/utils/clsx-helper";
+    import { cn } from "$lib/utils/shadcn-helper";
 
     let {
         ref = $bindable(null),
