+++ src/lib/components/primitives/dropdown-menu/dropdown-menu-checkbox-item.svelte
@@ -2,7 +2,7 @@
     import { DropdownMenu as DropdownMenuPrimitive, type WithoutChildrenOrChild } from "bits-ui";
     import Check from "lucide-svelte/icons/check";
     import Minus from "lucide-svelte/icons/minus";
-    import { cn } from "$lib/utils/clsx-helper";
+    import { cn } from "$lib/utils/shadcn-helper";
     import type { Snippet } from "svelte";
 
     let {
