+++ src/lib/components/primitives/dropdown-menu/dropdown-menu-label.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { cn } from "$lib/utils/clsx-helper";
+    import { cn } from "$lib/utils/shadcn-helper";
     import { type WithElementRef } from "bits-ui";
     import type { HTMLAttributes } from "svelte/elements";
