+++ src/lib/components/primitives/dropdown-menu/dropdown-menu-shortcut.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import type { HTMLAttributes } from "svelte/elements";
     import { type WithElementRef } from "bits-ui";
-    import { cn } from "$lib/utils/clsx-helper";
+    import { cn } from "$lib/utils/shadcn-helper";
 
     let {
         ref = $bindable(null),
