+++ src/lib/components/primitives/skeleton/skeleton.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import type { WithElementRef, WithoutChildren } from "bits-ui";
     import type { HTMLAttributes } from "svelte/elements";
-    import { cn } from "$lib/utils/clsx-helper";
+    import { cn } from "$lib/utils/shadcn-helper";
 
     let {
         ref = $bindable(null),
