+++ src/lib/components/primitives/tabs/tabs-trigger.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { Tabs as TabsPrimitive } from "bits-ui";
-    import { cn } from "$lib/utils/clsx-helper";
+    import { cn } from "$lib/utils/shadcn-helper";
 
     let {
         ref = $bindable(null),
