+++ src/routes/+page.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import SimpleNavigationBar from "$lib/components/composites/navigation-bar/SimpleNavigationBar.svelte";
-    import NamedList from "$lib/components/composites/utils/NamedList.svelte";
+    import NamedList from "$lib/components/composites/list/NamedList.svelte";
     import ProjectListEntry from "$lib/components/composites/project-components/ProjectListEntry.svelte";
     import ProjectListEntrySkeleton from "$lib/components/composites/project-components/ProjectListEntrySkeleton.svelte";
