+++ tests/integration/list/ExampleListItemSkeleton.svelte
