+++ tests/integration/list/ExampleSnippets.svelte
